refactor: adj path to fix runtime dep error
Ref. metriport/metriport-internal#2840
