+++ packages/api/src/external/ehr/athenahealth/command/sync-patient.ts
@@ -1,4 +1,4 @@
-import { isAthenaCustomFieldsEnabledForCx } from "@metriport/core/command/feature-flags";
+import { isAthenaCustomFieldsEnabledForCx } from "@metriport/core/command/feature-flags/domain-ffs";
 import AthenaHealthApi from "@metriport/core/external/ehr/athenahealth/index";
 import { processAsyncError } from "@metriport/core/util/error/shared";
 import { BadRequestError } from "@metriport/shared";
